Begin to decouple message queue consumers from the queue itself
