+++ packages/pds/src/db/message-queue/index.ts
@@ -1,36 +1,18 @@
 import Database from '..'
-import {
-  AddMember,
-  AddUpvote,
-  CreateNotification,
-  DeleteNotifications,
-  Message,
-  RemoveMember,
-  RemoveUpvote,
-} from './messages'
-import { sql } from 'kysely'
 import { dbLogger as log } from '../../logger'
-import { AuthStore } from '@atproto/auth'
-import * as repo from '../../repo'
-import * as lexicons from '../../lexicon/lexicons'
-import { TID } from '@atproto/common'
+import { Listener, BaseMessage as Msg } from '../../stream/types'
 import { MessageQueue } from '../types'
-
-type GetAuthStoreFn = (did: string) => AuthStore
 
 export class SqlMessageQueue implements MessageQueue {
   private cursorExists = false
   private ensureCaughtUpTimeout: ReturnType<typeof setTimeout> | undefined
+  private listeners: Map<string, Listener[]> = new Map()
 
-  constructor(
-    private name: string,
-    private db: Database,
-    private getAuthStore: GetAuthStoreFn,
-  ) {
+  constructor(private name: string, private db: Database) {
     this.ensureCaughtUp()
   }
 
-  async send(tx: Database, messages: Message | Message[]): Promise<void> {
+  async send(tx: Database, messages: Msg | Msg[]): Promise<void> {
     const msgArray = Array.isArray(messages) ? messages : [messages]
     if (msgArray.length === 0) return
     const now = new Date().toISOString()
@@ -45,6 +27,12 @@ export class SqlMessageQueue implements MessageQueue {
         log.error({ err }, 'error processing message')
       })
     }
+  }
+
+  listen<T extends string, M extends Msg<T>>(topic: T, listener: Listener<M>) {
+    const listeners = this.listeners.get(topic) ?? []
+    listeners.push(listener as Listener) // @TODO avoid upcast
+    this.listeners.set(topic, listeners)
   }
 
   private async ensureCursor(): Promise<void> {
@@ -90,7 +78,7 @@ export class SqlMessageQueue implements MessageQueue {
       if (res.length === 0) return
 
       for (const row of res) {
-        const message: Message = JSON.parse(row.message)
+        const message: Msg = JSON.parse(row.message)
         await this.handleMessage(dbTxn, message)
       }
       const nextCursor = res[res.length - 1].id + 1
@@ -122,7 +110,7 @@ export class SqlMessageQueue implements MessageQueue {
       // all caught up
       if (!res) return
 
-      const message: Message = JSON.parse(res.message)
+      const message: Msg = JSON.parse(res.message)
       await this.handleMessage(dbTxn, message)
 
       const nextCursor = res.id + 1
@@ -135,195 +123,18 @@ export class SqlMessageQueue implements MessageQueue {
     })
   }
 
-  private async handleMessage(db: Database, message: Message) {
-    try {
-      switch (message.type) {
-        case 'add_member':
-          return this.handleAddMember(db, message)
-        case 'remove_member':
-          return this.handleRemoveMember(db, message)
-        case 'add_upvote':
-          return this.handleAddUpvote(db, message)
-        case 'remove_upvote':
-          return this.handleRemoveUpvote(db, message)
-        case 'create_notification':
-          return this.handleCreateNotification(db, message)
-        case 'delete_notifications':
-          return this.handleDeleteNotifications(db, message)
+  private async handleMessage(db: Database, message: Msg) {
+    const listeners = this.listeners.get(message.type)
+    if (!listeners?.length) {
+      return log.error({ message }, `no listeners for event: ${message.type}`)
+    }
+    for (const listener of listeners) {
+      try {
+        await listener({ db, message }) // @TODO returned side effects
+      } catch (err) {
+        log.error({ message, err }, `unable to handle event: ${message.type}`)
       }
-    } catch (err) {
-      log.error({ message, err }, 'unable to handle event')
     }
-  }
-
-  // Reducers
-  // -------------
-  private async handleAddMember(
-    db: Database,
-    message: AddMember,
-  ): Promise<void> {
-    const res = await db.db
-      .updateTable('scene_member_count')
-      .set({ count: sql`count + 1` })
-      .where('did', '=', message.scene)
-      .returningAll()
-      .executeTakeFirst()
-    if (!res) {
-      await db.db
-        .insertInto('scene_member_count')
-        .values({ did: message.scene, count: 1 })
-        .execute()
-    }
-  }
-
-  private async handleRemoveMember(
-    db: Database,
-    message: RemoveMember,
-  ): Promise<void> {
-    await db.db
-      .updateTable('scene_member_count')
-      .set({ count: sql`count - 1` })
-      .where('did', '=', message.scene)
-      .returning(['did', 'count'])
-      .execute()
-  }
-
-  private async handleAddUpvote(
-    db: Database,
-    message: AddUpvote,
-  ): Promise<void> {
-    const userScenes = await db.getScenesForUser(message.user)
-    if (userScenes.length < 1) return
-    const updated = await db.db
-      .updateTable('scene_votes_on_post')
-      .set({ count: sql`count + 1` })
-      .where('subject', '=', message.subject)
-      .where('did', 'in', userScenes)
-      .returningAll()
-      .execute()
-    const updatedIds = updated.map((row) => row.did)
-    const toInsert = userScenes.filter((scene) => updatedIds.indexOf(scene) < 0)
-    if (toInsert.length > 0) {
-      await db.db
-        .insertInto('scene_votes_on_post')
-        .values(
-          toInsert.map((scene) => ({
-            did: scene,
-            subject: message.subject,
-            count: 1,
-            postedTrending: 0 as 0 | 1,
-          })),
-        )
-        .execute()
-    }
-    await this.checkTrending(db, userScenes, message.subject)
-  }
-
-  private async handleRemoveUpvote(
-    db: Database,
-    message: RemoveUpvote,
-  ): Promise<void> {
-    const userScenes = await db.getScenesForUser(message.user)
-    if (userScenes.length === 0) return
-    await db.db
-      .updateTable('scene_votes_on_post')
-      .set({ count: sql`count - 1` })
-      .where('subject', '=', message.subject)
-      .where('did', 'in', userScenes)
-      .returning(['did', 'count'])
-      .execute()
-  }
-
-  private async handleCreateNotification(
-    db: Database,
-    message: CreateNotification,
-  ): Promise<void> {
-    await db.db
-      .insertInto('user_notification')
-      .values({
-        userDid: message.userDid,
-        recordUri: message.recordUri,
-        recordCid: message.recordCid,
-        author: message.author,
-        reason: message.reason,
-        reasonSubject: message.reasonSubject,
-        indexedAt: new Date().toISOString(),
-      })
-      .returningAll()
-      .execute()
-  }
-
-  private async handleDeleteNotifications(
-    db: Database,
-    message: DeleteNotifications,
-  ) {
-    await db.db
-      .deleteFrom('user_notification')
-      .where('recordUri', '=', message.recordUri)
-      .execute()
-  }
-
-  // Side effects
-  // -------------
-  private async checkTrending(
-    db: Database,
-    scenes: string[],
-    subject: string,
-  ): Promise<void> {
-    if (scenes.length === 0) return
-    const state = await db.db
-      .selectFrom('scene_member_count as members')
-      .innerJoin('scene_votes_on_post as votes', 'votes.did', 'members.did')
-      .innerJoin('post', 'post.uri', 'votes.subject')
-      .where('members.did', 'in', scenes)
-      .where('votes.subject', '=', subject)
-      .select([
-        'members.did as did',
-        'members.count as memberCount',
-        'votes.subject as subject',
-        'post.cid as subjectCid',
-        'votes.count as voteCount',
-        'votes.postedTrending as postedTrending',
-      ])
-      .execute()
-
-    const now = new Date().toISOString()
-    await Promise.all(
-      state.map(async (scene) => {
-        if (scene.postedTrending) return
-        const ratio =
-          scene.memberCount !== 0 ? scene.voteCount / scene.memberCount : 0
-        const shouldTrend = scene.voteCount > 1 && ratio >= 0.2
-        if (!shouldTrend) return
-
-        // this is a "threshold vote" that makes the post trend
-        const sceneAuth = this.getAuthStore(scene.did)
-        const writes = await repo.prepareWrites(scene.did, {
-          action: 'create',
-          collection: lexicons.ids.AppBskyFeedTrend,
-          rkey: TID.nextStr(),
-          value: {
-            subject: {
-              uri: scene.subject,
-              cid: scene.subjectCid,
-            },
-            createdAt: now,
-          },
-        })
-        const setTrendPosted = db.db
-          .updateTable('scene_votes_on_post')
-          .set({ postedTrending: 1 })
-          .where('did', '=', scene.did)
-          .where('subject', '=', scene.subject)
-          .execute()
-
-        await Promise.all([
-          repo.writeToRepo(db, scene.did, sceneAuth, writes, now),
-          repo.indexWrites(db, writes, now),
-          setTrendPosted,
-        ])
-      }),
-    )
   }
 }
