+++ packages/pds/src/db/record-processor.ts
@@ -3,7 +3,7 @@ import * as common from '@atproto/common'
 import { Kysely } from 'kysely'
 import { CID } from 'multiformats/cid'
 import { DatabaseSchema } from './database-schema'
-import { Message } from './message-queue/messages'
+import { Message } from '../stream/messages'
 import * as schemas from './schemas'
 
 type RecordProcessorParams<T, S> = {
