+++ packages/pds/src/db/records/assertion.ts
@@ -4,8 +4,8 @@ import { CID } from 'multiformats/cid'
 import * as Assertion from '../../lexicon/types/app/bsky/graph/assertion'
 import { Assertion as IndexedAssertion } from '../tables/assertion'
 import * as schemas from '../schemas'
-import * as messages from '../message-queue/messages'
-import { Message } from '../message-queue/messages'
+import * as messages from '../../stream/messages'
+import { Message } from '../../stream/messages'
 import DatabaseSchema from '../database-schema'
 import RecordProcessor from '../record-processor'
