+++ packages/pds/src/db/records/confirmation.ts
@@ -4,8 +4,8 @@ import { CID } from 'multiformats/cid'
 import * as Confirmation from '../../lexicon/types/app/bsky/graph/confirmation'
 import { Assertion as IndexedAssertion } from '../tables/assertion'
 import * as schemas from '../schemas'
-import * as messages from '../message-queue/messages'
-import { Message } from '../message-queue/messages'
+import * as messages from '../../stream/messages'
+import { Message } from '../../stream/messages'
 import { APP_BSKY_GRAPH } from '../../lexicon'
 import DatabaseSchema from '../database-schema'
 import RecordProcessor from '../record-processor'
