+++ packages/pds/src/db/records/follow.ts
@@ -4,8 +4,8 @@ import { CID } from 'multiformats/cid'
 import * as Follow from '../../lexicon/types/app/bsky/graph/follow'
 import { Follow as IndexedFollow } from '../tables/follow'
 import * as schemas from '../schemas'
-import * as messages from '../message-queue/messages'
-import { Message } from '../message-queue/messages'
+import * as messages from '../../stream/messages'
+import { Message } from '../../stream/messages'
 import DatabaseSchema from '../database-schema'
 import RecordProcessor from '../record-processor'
