+++ packages/pds/src/db/records/post.ts
@@ -4,8 +4,8 @@ import { CID } from 'multiformats/cid'
 import * as Post from '../../lexicon/types/app/bsky/feed/post'
 import * as PostTables from '../tables/post'
 import * as schemas from '../schemas'
-import * as messages from '../message-queue/messages'
-import { Message } from '../message-queue/messages'
+import * as messages from '../../stream/messages'
+import { Message } from '../../stream/messages'
 import DatabaseSchema from '../database-schema'
 import RecordProcessor from '../record-processor'
