+++ packages/pds/src/db/records/repost.ts
@@ -4,8 +4,8 @@ import * as Repost from '../../lexicon/types/app/bsky/feed/repost'
 import { Repost as IndexedRepost } from '../tables/repost'
 import * as schemas from '../schemas'
 import { CID } from 'multiformats/cid'
-import * as messages from '../message-queue/messages'
-import { Message } from '../message-queue/messages'
+import * as messages from '../../stream/messages'
+import { Message } from '../../stream/messages'
 import { DatabaseSchema } from '../database-schema'
 import RecordProcessor from '../record-processor'
