+++ packages/pds/src/db/records/trend.ts
@@ -4,8 +4,8 @@ import * as Trend from '../../lexicon/types/app/bsky/feed/trend'
 import { Trend as IndexedTrend } from '../tables/trend'
 import * as schemas from '../schemas'
 import { CID } from 'multiformats/cid'
-import * as messages from '../message-queue/messages'
-import { Message } from '../message-queue/messages'
+import * as messages from '../../stream/messages'
+import { Message } from '../../stream/messages'
 import DatabaseSchema from '../database-schema'
 import RecordProcessor from '../record-processor'
