+++ packages/pds/src/db/records/vote.ts
@@ -4,8 +4,8 @@ import { CID } from 'multiformats/cid'
 import * as Vote from '../../lexicon/types/app/bsky/feed/vote'
 import { Vote as IndexedVote } from '../tables/vote'
 import * as schemas from '../schemas'
-import * as messages from '../message-queue/messages'
-import { Message } from '../message-queue/messages'
+import * as messages from '../../stream/messages'
+import { Message } from '../../stream/messages'
 import { DatabaseSchema } from '../database-schema'
 import RecordProcessor from '../record-processor'
