+++ packages/pds/src/db/types.ts
@@ -1,7 +1,8 @@
 import { AtUri } from '@atproto/uri'
 import { CID } from 'multiformats/cid'
 import { DynamicReferenceBuilder } from 'kysely/dist/cjs/dynamic/dynamic-reference-builder'
-import { Message } from './message-queue/messages'
+import { BaseMessage, Listener } from '../stream/types'
+import { Message } from '../stream/messages'
 import Database from '.'
 
 export type DbRecordPlugin<T> = {
@@ -20,7 +21,11 @@ export type DbRecordPlugin<T> = {
 export type Ref = DynamicReferenceBuilder<any>
 
 export interface MessageQueue {
-  send(tx: Database, message: Message | Message[]): Promise<void>
+  send(tx: Database, message: BaseMessage | BaseMessage[]): Promise<void>
+  listen<T extends string, M extends BaseMessage<T>>(
+    topic: T,
+    listener: Listener<M>,
+  ): void
   processNext(): Promise<void>
   processAll(): Promise<void>
   destroy(): void
